Extract form interests controller from announcement form js
