+++ assets/js/announcement-form.js
@@ -1,8 +1,4 @@
 import { markedWithSyntax } from './syntax-highlighting';
-import { updateRecipientsPreview } from './recipients-preview';
-import 'selectize';
-
-const DELIMITER = ',';
 
 export default class {
   constructor() {
@@ -11,7 +7,6 @@ export default class {
 
     this.watchTitle();
     this.watchBody();
-    this.setupInterestsSelect();
     this.clearLocalStorageOnSubmit();
   }
 
@@ -69,36 +64,6 @@ export default class {
       $('[data-role=markdown-preview]').html(markdown);
     }
   };
-
-  setupInterestsSelect() {
-    const interests = $('#announcement_interests');
-
-    if (interests.length !== 0) {
-      if (interests.val() === '') {
-        const localStorageValue = localStorage.getItem('interests');
-        interests.val(localStorageValue);
-        updateRecipientsPreview(localStorageValue);
-      }
-
-      interests.selectize({
-        delimiter: DELIMITER,
-        persist: false,
-        create: function(name) {
-          return { name };
-        },
-        valueField: 'name',
-        labelField: 'name',
-        searchField: 'name',
-        options: window.INTERESTS_NAMES,
-        onChange: (value) => {
-          if (!this._isEditing) {
-            localStorage.setItem('interests', value);
-          }
-          updateRecipientsPreview(value);
-        },
-      });
-    }
-  }
 
   clearLocalStorageOnSubmit() {
     if (!this._isEditing) {
